Moved components to Components folder on example project
