+++ example/JJMasterData.Web.Example/Pages/Components/JJAlertExample.cshtml
