+++ example/JJMasterData.Web.Example/Pages/Components/JJTitleExample.cshtml.cs
@@ -1,8 +1,8 @@
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace JJMasterData.Web.Example.Pages;
+namespace JJMasterData.Web.Example.Pages.Components;
 
-public class JJAlertExample : PageModel
+public class JJTitleExample : PageModel
 {
     public void OnGet()
     {
