+++ example/JJMasterData.Web.Example/Pages/Shared/_Layout.cshtml
@@ -23,9 +23,9 @@
     <ul class="menu">
         <li><a asp-page="/Index">@Localizer["Intro"]</a></li>
         <li><a href="/pt-br/DataDictionary">@Localizer["Data Dictionary"]</a></li>
-        <li><a asp-page="/JJFormViewExample">JJFormView</a></li>
-        <li><a asp-page="/JJTitleExample">JJTitle</a></li>
-        <li><a asp-page="/JJAlertExample">JJAlert</a></li>
+        <li><a asp-page="/Components/JJFormViewExample">JJFormView</a></li>
+        <li><a asp-page="/Components/JJTitleExample">JJTitle</a></li>
+        <li><a asp-page="/Components/JJAlertExample">JJAlert</a></li>
     </ul>
 </header>
 <main>
